Moved NarrativeCase to its own file
