+++ unity-ggjj/Assets/Scripts/Menu/CaseSelectMenu.cs
@@ -44,11 +44,3 @@ public class CaseSelectMenu : MonoBehaviour
         }
     }
 }
-
-[Serializable]
-public struct NarrativeCase
-{
-    [field: SerializeField] public string Name { get; private set; }
-    [field: SerializeField] public Sprite PreviewImage { get; private set; }
-    [field: SerializeField] public TextAsset[] Chapters { get; private set; }
-}
