+++ unity-ggjj/Assets/Scripts/Menu/NarrativeCase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 027da1b7faac8428d8a760e3e8f66131
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
